add leaderboard back

+++ src/main/java/com/school21/Tic_Tac_Toe/datasource/model/UserRatioProjection.java
@@ -4,5 +4,5 @@ import java.util.UUID;
 
 public interface UserRatioProjection {
     UUID getUserId();
-    Float getWinRatio();
+    Double getWinRatio();
 }
