+++ src/main/java/com/school21/Tic_Tac_Toe/domain/model/stats/UserRatioModel.java
@@ -11,5 +11,5 @@ import java.util.UUID;
 @NoArgsConstructor
 public class UserRatioModel {
     private UUID userId;
-    private float winRatio;
+    private double winRatio;
 }
